codegenie2/requirements.txt
streamlit
e2b_code_interpreter
litellm
tavily-python